Update readme.txt
Hi Patrick, 

I'm Russ and a Sales Engineer at GoSquared.  We've been made aware by a few newer users now that a warning notice is appearing indicating that the plugin has not been tested on the more current versions of Wordpress.

We've tested on 4.82 so just wanted to reflect this on the Readme.txt 

Would it be possible to publish again?